Move passes from `Passes.td` -> `LLVMGPUPasses.td`.
